requirements.txt
Flask==3.0.0
openai==0.28.0
openpyxl==3.1.2
pandas==2.2.2
python-dotenv==1.0.0
flask_httpauth==4.8.0
werkzeug==2.0.0
